requirements.txt
aiogram==2.11.2
aiohttp>=3.6.2
aioredis==1.3.1
python-dotenv~=0.17.0




pip~=21.0.1
attrs~=20.3.0
pytz~=2021.1
idna~=2.10
multidict~=5.1.0
Babel~=2.9.0
certifi~=2020.12.5
yarl~=1.6.3
chardet~=4.0.0
hiredis~=2.0.0
setuptools~=54.2.0
pyqiwi~=0.2.0